Less strict hierarchy for contacts check
Checks always all the contact fields first name, email, affiliation and
orcid if there is a last name. If all checks pass, `True` gets appended
to the response list. If last name is missing the checks for other
fields are still skipped but a string indicating that last name is
missing gets appended to the response.